+++ apps/sponsor/lib/sponsor/purchases/purchase.ex
@@ -18,7 +18,7 @@ defmodule Tankste.Sponsor.Purchases.Purchase do
     struct
     |> cast(params, [:device_id, :product, :provider, :type])
     |> validate_required([:device_id, :product, :provider, :type])
-    |> validate_inclusion(:product, ~w(sponsor_single_1 sponsor_single_2 sponsor_single_10 sponsor_subscription_monthly_1 sponsor_subscription_monthly_2 sponsor_subscription_yearly_12))
+    |> validate_inclusion(:product, ~w(sponsor_single_1 sponsor_single_2 sponsor_single_5 sponsor_single_10 sponsor_single_20 sponsor_single_50 sponsor_subscription_monthly_1 sponsor_subscription_monthly_2 sponsor_subscription_yearly_12))
     |> validate_inclusion(:provider, ~w(apple_store play_store))
     |> validate_inclusion(:type, ~w(single subscription))
   end